+++ internal/provider/provider.go
@@ -306,7 +306,6 @@ func (p *HostProvider) Resources(ctx context.Context) []func() resource.Resource
 	return []func() resource.Resource{
 		NewDNFPackageResource,
 		NewPacmanPackageResource,
-		NewAURHelperResource,
 		NewAURPackageResource,
 		NewBrewPackageResource,
 		NewHostDirResource,